+++ Options/CHANGELOG.lua
@@ -8,6 +8,7 @@ TMW.CHANGELOG = [==[
 
 ### Bug Fixes
 * Fix #1742 - Errors related to improper escaping of user input for the suggestion list.
+* Fix #1755 - Swing Timer conditions with durations other than zero seconds were not triggering updates at the proper moment.
 
 ## v8.7.2
 ### Bug Fixes